format with latest elm-format and limit traversals
By only traversing the entire list of quads once, we save a fair amount of work.
Using a single `foldl` and passing both the "incremental offset" as well as the
resulting (flat) list of quads around, we can prevent needing subsequent maps.

Using a dirty trick in `circularGet : Int -> Array a -> a`, we can also prevent
a whole bunch of `Maybe` logic, which should also impact runtime, at the cost of
potentially going into an endless loop when passing a negative index.